python/ecuconnect_tool/pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.5", "pybind11>=2.11"]
build-backend = "scikit_build_core.build"

[project]
name = "ecuconnect-tool"
version = "0.1.0"
description = "Python ECUconnect CLI and CANyonero protocol bindings"
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "ECUconnect" }]
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "python-can>=4.2.2",
]

[project.scripts]
ecuconnect-tool = "ecuconnect_tool.cli:app"

[tool.scikit-build]
cmake.source-dir = "."
wheel.packages = ["ecuconnect_tool"]
